Restructure front directory, alter gitignore
Scope: Front end

-Restructure front dir. js own dir, css own dir.
-Alter gitignore to ignore node_modules.
