+++ cezaro/.gitignore
@@ -1,26 +1,53 @@
 HELP.md
+
 /target/
+
 !.mvn/wrapper/maven-wrapper.jar
 
+
+
 ### STS ###
+
 .apt_generated
+
 .classpath
+
 .factorypath
+
 .project
+
 .settings
+
 .springBeans
+
 .sts4-cache
 
+
+
 ### IntelliJ IDEA ###
+
 .idea
+
 *.iws
+
 *.iml
+
 *.ipr
 
+
+
 ### NetBeans ###
+
 /nbproject/private/
+
 /nbbuild/
+
 /dist/
+
 /nbdist/
+
 /.nb-gradle/
+
 /build/
+
+/front/node_modules/
