+++ cezaro/front/src/components/Post.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
-import './Post.css';
+
+require('../styles/Post.css');
 
 class Post extends Component {
