+++ cezaro/front/src/config/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+// Settings configured here will be merged into the final config object.
+export default {
+    api: 'http://localhost:8080'
+}
