+++ cezaro/front/src/styles/LeftPanel.css
