+++ cezaro/front/src/styles/Post.css
