+++ cezaro/front/src/styles/SearchComponent.css
